+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+v#.#.# (MMMM YYYY)
+  - Add support for tests that start with `ContainerBlockElement` tag rather than `Paragraph`
+
 v4.9.0 (June 2023)
   - Parse inline code, not just code blocks
   - Wrap ciphers in the `ssl-weak-message-authentication-code-algorithms` finding